Move logic for `ensure-online`
