+++ source/util.ts
@@ -33,20 +33,3 @@ export function showRestartDialog(message: string): void {
 		}
 	);
 }
-
-export function showRetryDialog(message: string): void {
-	dialog.showMessageBox(
-		{
-			message,
-			detail: `Do you want to wait?`,
-			buttons: ['Wait', 'Quit'],
-			defaultId: 0,
-			cancelId: 1
-		},
-		response => {
-			if (response === 1) {
-				app.quit();
-			}
-		}
-	);
-}
